fix: Migrate arena-hard card to lmarena-ai/arena-hard-viewer and fix WeightedWinRateCorrelation metric
The old lmsys/arena-hard-browser HF space is no longer available. This
migrates to the replacement space lmarena-ai/arena-hard-viewer with
adapted processing steps for its different data format (flat prompt
field, messages-based answers, uid instead of question_id).

Also fixes a bug in WeightedWinRateCorrelation where pd.DataFrame
columns initialized as object dtype caused scipy pearsonr to fail
with newer numpy/scipy versions.

+++ src/unitxt/metrics.py
@@ -1929,10 +1929,12 @@ class WeightedWinRateCorrelation(GlobalMetric):
             pred_df_win_rate, ref_df_win_rate, on="model", suffixes=("_pred", "_ref")
         )
         pearson_corr, _ = pearsonr(
-            merged_df["win_rate_pred"], merged_df["win_rate_ref"]
+            merged_df["win_rate_pred"].astype(float),
+            merged_df["win_rate_ref"].astype(float),
         )
         spearman_corr, _ = spearmanr(
-            merged_df["win_rate_pred"], merged_df["win_rate_ref"]
+            merged_df["win_rate_pred"].astype(float),
+            merged_df["win_rate_ref"].astype(float),
         )
 
         return {"pearson_corr": pearson_corr, "spearman_corr": spearman_corr}
